Rename class in Oak branch
