+++ AccountList.py
@@ -1,4 +1,4 @@
-from user import User
+from User import User
 
 class AccountList:
     def __init__(self):
