+++ Owner.py
@@ -1,4 +1,4 @@
-from user import User
+from User import User
 
 class Owner(User):
     def __init__(self,name,lastname,email,user_name,password,user_phone,verified):
