--- tempCodeRunnerFile.py
@@ -1,3 +1,0 @@
-print(Dorcat.search_fac_dor("pets"))
-print(Dorcat.search_fac_dor("smoking"))
-print(Dorcat.search_maxmin_price(4000,6600))
